// src/app/models/prediccion.ts
export interface Prediccion{
    idPredicciones?: number;
    fecha_prediccion?: string;
    jornada_prediccion?: string;
    numeros?: {
        1?: {
            numero?: number
            probabilidad?: string
        }
        2?: {
            numero?: number
            probabilidad?: string
        }
        3?: {
            numero?: number
            probabilidad?: string
        }
        4?: {
            numero?: number
            probabilidad?: string
        }
        5?: {
            numero?: number
            probabilidad?: string
        }
        6?: {
            numero?: number
            probabilidad?: string
        }
        7?: {
            numero?: number
            probabilidad?: string
        }
        8?: {
            numero?: number
            probabilidad?: string
        }
        9?: {
            numero?: number
            probabilidad?: string
        }
        10?: {
            numero?: number
            probabilidad?: string
        }
    }
}